Allow configuring active plugins
Fixes #291
