+++ CHANGES/291.feature
@@ -1,0 +1,1 @@
+Add `plugins` configuration option to select which plugins to load.
